move artifact/cos to storage
