+++ storage/cos/go.mod
@@ -1,0 +1,21 @@
+module trpc.group/trpc-go/trpc-agent-go/storage/cos
+
+replace trpc.group/trpc-go/trpc-agent-go => ../../
+
+go 1.24.1
+
+require (
+	github.com/stretchr/testify v1.11.1
+	github.com/tencentyun/cos-go-sdk-v5 v0.7.69
+	trpc.group/trpc-go/trpc-agent-go v0.0.0-00010101000000-000000000000
+)
+
+require (
+	github.com/clbanning/mxj v1.8.4 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/go-querystring v1.0.0 // indirect
+	github.com/mitchellh/mapstructure v1.4.3 // indirect
+	github.com/mozillazg/go-httpheader v0.2.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
