+++ storage/cos/option.go
