+++ storage/cos/service_test.go
